requirements.txt
Django
django-cors-headers
django-debug-toolbar
django-filter
django-rest-registration
djangorestframework
drf-yasg
django-ckeditor
python-decouple
python-dotenv
coreapi
Pillow
django-cleanup

